+++ src/index.ts
@@ -1,5 +1,6 @@
 import SetOf from "./classes/SetOf.ts";
-import AreArraySetsEqual from "./utils/AreArraySetsEqual.ts";
+import areArraySetsEqual from "./utils/areArraySetsEqual.ts";
+import arraySetDifference from "./utils/arraySetDifference.ts";
 import isPartOf from "./utils/isPartOf.ts";
 import powerSet from "./utils/powerSet.ts";
 import arrayToSet from "./utils/utils.ts";
@@ -9,6 +10,7 @@ export default SetOf;
 export {
     powerSet,
     isPartOf,
-    AreArraySetsEqual,
+    areArraySetsEqual,
     arrayToSet,
+    arraySetDifference,
 };