+++ src/utils/AreArraySetsEqual_test.ts
@@ -1,11 +1,11 @@
 import { assertEquals } from "@std/assert/equals";
-import AreArraySetsEqual from "./AreArraySetsEqual.ts";
+import areArraySetsEqual from "./areArraySetsEqual.ts";
 
 Deno.test( "[1, 2] is equal to [1,2]", () => {
     const a = [1, 2]
     const b = [1, 2]   
 
-    assertEquals( AreArraySetsEqual(a, b), true )
+    assertEquals( areArraySetsEqual(a, b), true )
 })
 
 
@@ -13,7 +13,7 @@ Deno.test( "[1, 2] is equal to [2, 1]", () => {
     const a = [1, 2]
     const b = [2, 1]   
 
-    assertEquals( AreArraySetsEqual(a, b), true )
+    assertEquals( areArraySetsEqual(a, b), true )
 })
 
 
@@ -21,7 +21,7 @@ Deno.test( "[1, 2 , [1]] is equal to [2, 1, [1]]", () => {
     const a = [1, 2, [1]]
     const b = [2, 1, [1]]   
 
-    assertEquals( AreArraySetsEqual(a, b), true )
+    assertEquals( areArraySetsEqual(a, b), true )
 })
 
 
@@ -29,7 +29,7 @@ Deno.test( "[1, 2 , [1, [2]]] is equal to [2, 1, [1, [2]]]", () => {
     const a = [ 1, 2, [1, [2]] ]
     const b = [ 2, 1, [1, [2]] ]   
 
-    assertEquals( AreArraySetsEqual(a, b), true )
+    assertEquals( areArraySetsEqual(a, b), true )
 })
 
 
@@ -38,22 +38,22 @@ Deno.test( "[ [1, [2 , [3] ]], 1, 2 ]  is equal to [ 1, 2, [1, [2 , [3] ]] ]", (
     const a = [ [1, [2 , [3] ]], 1, 2 ] 
     const b = [ 1, 2, [1, [2 , [3] ]] ]
 
-    assertEquals( AreArraySetsEqual(a, b), true )
+    assertEquals( areArraySetsEqual(a, b), true )
 })
 
 
 Deno.test(" [[4]] is not equal to  [ [ [4] ] ]  ", () => {
     const a = [[4]];
     const b = [ [ [4] ] ];
-    assertEquals( AreArraySetsEqual(a,b), false )
+    assertEquals( areArraySetsEqual(a,b), false )
 })
 
 
 Deno.test(" [] is equal to [] ", () => {
-    assertEquals( AreArraySetsEqual([],[]), true )
+    assertEquals( areArraySetsEqual([],[]), true )
 
 })
 
 Deno.test(" [] is not equal to [null] ", () => {
-    assertEquals( AreArraySetsEqual([],[null]), false )
+    assertEquals( areArraySetsEqual([],[null]), false )
 })